+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented here. The format is based on [Keep a Changelog][keep-a-changelog].
 
-## Unreleased
+## 0.4.1 (2018-05-15)
 
 -   Added Duration -> arbitrary format pipe.
 -   Added longest/shortest Duration pipes.